Refactor create/edit views
